refactor(PostHeader): 게시물 삭제 로직 수정

+++ src/pages/PostDetailPage/component/PostHeader.jsx
@@ -1,24 +1,29 @@
 import { useNavigate } from "react-router-dom";
-
+import { requestCsrfAPIJsonResponse } from "../../../api/csrf";
 export const PostHeader = ({user, post, isAuthor}) =>{
     const navigate = useNavigate();
 
     async function handleDelete() {
-        try{
-            const response = await fetch(`http://localhost:8080/posts/${postId}`, {
-                method: 'DELETE',
-                credentials:"include",
-                headers: {
-                    'Content-Type': 'application/json',
-                    [csrf.headerName] : csrf.token
-                },
-            });
+        if(window.confirm('게시글을 삭제하시겠습니까?')){
+            const csrf = await requestCsrfAPIJsonResponse();
+            
+            try{
+                const response = await fetch(`http://localhost:8080/posts/${post.id}`, {
+                    method: 'DELETE',
+                    credentials:"include",
+                    headers: {
+                        'Content-Type': 'application/json',
+                        [csrf.headerName] : csrf.token
+                    },
+                });
 
-            if (!response.ok) {
-                throw new Error('게시물 상세 조회 실패');
+                if (!response.ok) {
+                    throw new Error('게시물 상세 조회 실패');
+                }
+                navigate('/board')
+            }catch(error){
+                console.error('boardView 오류 발생:', error);
             }
-        }catch(error){
-            console.error('boardView 오류 발생:', error);
         }
     }  
     return(
